extract hug contents strategy to separate file
